requirements.txt
appdirs==1.4.4
atomicwrites==1.4.1
dateparser==1.3.0
jsonpointer==3.0.0
protobuf==7.34.0
pyasn1-modules==0.4.2
pycryptodome==3.23.0
python-binance==1.0.35
tzlocal==5.3.1
websockets==16.0
wheel==0.46.3
binance-connector==3.12.0
websocket-client==1.9.0
python-dotenv==1.2.2
